+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add Abstract Class `Controller`
 * Trait: `CallMethod`
 * Helpers: `loadenv`, `env`
+* `e` helper: Add double-encode parameter, pass encoding & flags to `htmlspecialchars`
 
 ## [v0.0.02](https://github.com/parsa-mostafaie/pluslib/compare/v0.0.01...v0.0.02) - 8/26/2024